requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
pymupdf
pydantic
sqlalchemy
requests
numpy
pillow
opencv-python
easyocr
insightface
xgboost
torch
torchvision
psycopg2-binary
redis
